requirements/base.txt
# Django
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.9.0
django-configurations==2.2

# Channals
channels==2.1.6
channels-redis==2.3.2
daphne==2.2.4

# Http
requests==2.20.1

# Database
dj-database-url==0.5.0
psycopg2==2.7.5

# JSON
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.0

# Celery&Django-Celery-Beat
celery==4.3.0
django_celery_beat==1.4.0

# Redis
redis==3.2.0

